refactor(web): put predictive-text correction search-batcher in separate file
